Use only UTF-8 for fetching a message.
Currently the API lets the users fetch SMS using either UTF-8 or native
SMS-PDU format. SMS-PDU is not supported by the String() class (also its
more complicated than just a simple string encoding since it contains
recipient etc information). Besides, the library does not decode such
messages and leaves it to the user to handle them. Currently it decodes
it as the default character set on device which is incorrect behavior.
Hence in this change we ensure that the API can only be used with UTF-8
which makes the conversion of Strings consistent.

NOTE: It's OK to avoid supporting native charset since it is not
mandatory in MAPv12 spec.

Change-Id: Ic10a82ff69f848cb3f6e80353d3e63ce3910d306
(cherry picked from commit 4b4d3f8260c9b4476f900ee7b08ee66450624081)